+++ migrations/1677284495299_locations.ts
@@ -6,15 +6,18 @@ export const shorthands: ColumnDefinitions | undefined = undefined;
 export function up(pgm: MigrationBuilder): void {
   pgm.createTable('locations', {
     id: {
-      type: 'serial',
+      type: 'bigserial',
       primaryKey: true,
     },
     inscription_id: {
-      type: 'int',
+      type: 'bigint',
+    },
+    genesis_id: {
+      type: 'text',
       notNull: true,
     },
     block_height: {
-      type: 'int',
+      type: 'bigint',
       notNull: true,
     },
     block_hash: {
@@ -35,33 +38,17 @@ export function up(pgm: MigrationBuilder): void {
     offset: {
       type: 'numeric',
     },
+    prev_output: {
+      type: 'text',
+    },
+    prev_offset: {
+      type: 'numeric',
+    },
     value: {
       type: 'numeric',
     },
-    sat_ordinal: {
-      type: 'numeric',
-      notNull: true,
-    },
-    sat_rarity: {
-      type: 'text',
-      notNull: true,
-    },
-    sat_coinbase_height: {
-      type: 'int',
-      notNull: true,
-    },
     timestamp: {
       type: 'timestamptz',
-      notNull: true,
-    },
-    genesis: {
-      type: 'boolean',
-      default: true,
-      notNull: true,
-    },
-    current: {
-      type: 'boolean',
-      default: true,
       notNull: true,
     },
   });
@@ -70,17 +57,12 @@ export function up(pgm: MigrationBuilder): void {
     'locations_inscription_id_fk',
     'FOREIGN KEY(inscription_id) REFERENCES inscriptions(id) ON DELETE CASCADE'
   );
-  pgm.createConstraint(
-    'locations',
-    'locations_inscription_id_block_height_unique',
-    'UNIQUE(inscription_id, block_height)'
-  );
+  pgm.createConstraint('locations', 'locations_output_offset_unique', 'UNIQUE(output, "offset")');
+  pgm.createIndex('locations', ['inscription_id']);
+  pgm.createIndex('locations', ['genesis_id']);
   pgm.createIndex('locations', ['block_height']);
   pgm.createIndex('locations', ['block_hash']);
   pgm.createIndex('locations', ['address']);
-  pgm.createIndex('locations', ['output']);
-  pgm.createIndex('locations', ['sat_ordinal']);
-  pgm.createIndex('locations', ['sat_rarity']);
-  pgm.createIndex('locations', ['sat_coinbase_height']);
   pgm.createIndex('locations', ['timestamp']);
+  pgm.createIndex('locations', ['prev_output']);
 }