+++ migrations/1687785552000_inscriptions-per-block.ts
@@ -4,20 +4,26 @@ import { MigrationBuilder, ColumnDefinitions } from 'node-pg-migrate';
 export const shorthands: ColumnDefinitions | undefined = undefined;
 
 export function up(pgm: MigrationBuilder): void {
-  pgm.createTable('genesis', {
-    inscription_id: {
-      type: 'bigint',
-      notNull: true,
-    },
-    location_id: {
-      type: 'bigint',
-      notNull: true,
-    },
+  pgm.createTable('inscriptions_per_block', {
     block_height: {
       type: 'bigint',
+      primaryKey: true,
+    },
+    block_hash: {
+      type: 'text',
+      notNull: true,
+    },
+    inscription_count: {
+      type: 'bigint',
+      notNull: true,
+    },
+    inscription_count_accum: {
+      type: 'bigint',
+      notNull: true,
+    },
+    timestamp: {
+      type: 'timestamptz',
       notNull: true,
     },
   });
-  pgm.createConstraint('genesis', 'genesis_inscription_id_unique', 'UNIQUE(inscription_id)');
-  pgm.createIndex('genesis', ['location_id']);
 }