+++ src/pg/pg-store.ts
@@ -1,7 +1,7 @@
+import { BitcoinEvent, Payload } from '@hirosystems/chainhook-client';
 import { Order, OrderBy } from '../api/schemas';
 import { isProdEnv, normalizedHexString, parseSatPoint } from '../api/util/helpers';
 import { OrdinalSatoshi, SatoshiRarity } from '../api/util/ordinal-satoshi';
-import { ChainhookPayload } from '../chainhook/schemas';
 import { ENV } from '../env';
 import { logger } from '../logger';
 import { getIndexResultCountType, inscriptionContentToJson } from './helpers';
@@ -11,6 +11,8 @@ import { BasePgStore } from './postgres-tools/base-pg-store';
 import {
   DbFullyLocatedInscriptionResult,
   DbInscriptionContent,
+  DbInscriptionCountPerBlock,
+  DbInscriptionCountPerBlockFilters,
   DbInscriptionIndexFilters,
   DbInscriptionIndexOrder,
   DbInscriptionIndexPaging,
@@ -56,22 +58,23 @@ export class PgStore extends BasePgStore {
    * chain re-orgs and materialized view refreshes.
    * @param args - Apply/Rollback Chainhook events
    */
-  async updateInscriptions(payload: ChainhookPayload): Promise<void> {
+  async updateInscriptions(payload: Payload): Promise<void> {
+    let updatedBlockHeightMin = Infinity;
     await this.sqlWriteTransaction(async sql => {
-      for (const event of payload.rollback) {
+      for (const rollbackEvent of payload.rollback) {
+        const event = rollbackEvent as BitcoinEvent;
+        const block_height = event.block_identifier.index;
         for (const tx of event.transactions) {
           for (const operation of tx.metadata.ordinal_operations) {
             if (operation.inscription_revealed) {
               const number = operation.inscription_revealed.inscription_number;
               const genesis_id = operation.inscription_revealed.inscription_id;
-              await this.rollBackInscriptionGenesis({ genesis_id });
-              logger.info(`PgStore rollback reveal #${number} (${genesis_id})`);
+              await this.rollBackInscription({ genesis_id, number, block_height });
             }
             if (operation.cursed_inscription_revealed) {
               const number = operation.cursed_inscription_revealed.inscription_number;
               const genesis_id = operation.cursed_inscription_revealed.inscription_id;
-              await this.rollBackInscriptionGenesis({ genesis_id });
-              logger.info(`PgStore rollback cursed reveal #${number} (${genesis_id})`);
+              await this.rollBackInscription({ genesis_id, number, block_height });
             }
             if (operation.inscription_transferred) {
               const genesis_id = operation.inscription_transferred.inscription_id;
@@ -79,13 +82,14 @@ export class PgStore extends BasePgStore {
                 operation.inscription_transferred.satpoint_post_transfer
               );
               const output = `${satpoint.tx_id}:${satpoint.vout}`;
-              const id = await this.rollBackInscriptionTransfer({ genesis_id, output });
-              logger.info(`PgStore rollback transfer (${genesis_id}) ${output}`);
+              await this.rollBackLocation({ genesis_id, output, block_height });
             }
           }
         }
+        updatedBlockHeightMin = Math.min(updatedBlockHeightMin, event.block_identifier.index);
       }
-      for (const event of payload.apply) {
+      for (const applyEvent of payload.apply) {
+        const event = applyEvent as BitcoinEvent;
         const block_height = event.block_identifier.index;
         const block_hash = normalizedHexString(event.block_identifier.hash);
         for (const tx of event.transactions) {
@@ -95,7 +99,7 @@ export class PgStore extends BasePgStore {
               const reveal = operation.inscription_revealed;
               const satoshi = new OrdinalSatoshi(reveal.ordinal_number);
               const satpoint = parseSatPoint(reveal.satpoint_post_inscription);
-              await this.insertInscriptionGenesis({
+              await this.insertInscription({
                 inscription: {
                   genesis_id: reveal.inscription_id,
                   mime_type: reveal.content_type.split(';')[0],
@@ -105,6 +109,9 @@ export class PgStore extends BasePgStore {
                   content: reveal.content_bytes,
                   fee: reveal.inscription_fee.toString(),
                   curse_type: null,
+                  sat_ordinal: reveal.ordinal_number.toString(),
+                  sat_rarity: satoshi.rarity,
+                  sat_coinbase_height: satoshi.blockHeight,
                 },
                 location: {
                   block_hash,
@@ -118,20 +125,14 @@ export class PgStore extends BasePgStore {
                   prev_offset: null,
                   value: reveal.inscription_output_value.toString(),
                   timestamp: event.timestamp,
-                  sat_ordinal: reveal.ordinal_number.toString(),
-                  sat_rarity: satoshi.rarity,
-                  sat_coinbase_height: satoshi.blockHeight,
                 },
               });
-              logger.info(
-                `PgStore reveal #${reveal.inscription_number} (${reveal.inscription_id}) at block ${block_height}`
-              );
             }
             if (operation.cursed_inscription_revealed) {
               const reveal = operation.cursed_inscription_revealed;
               const satoshi = new OrdinalSatoshi(reveal.ordinal_number);
               const satpoint = parseSatPoint(reveal.satpoint_post_inscription);
-              await this.insertInscriptionGenesis({
+              await this.insertInscription({
                 inscription: {
                   genesis_id: reveal.inscription_id,
                   mime_type: reveal.content_type.split(';')[0],
@@ -141,6 +142,9 @@ export class PgStore extends BasePgStore {
                   content: reveal.content_bytes,
                   fee: reveal.inscription_fee.toString(),
                   curse_type: JSON.stringify(reveal.curse_type),
+                  sat_ordinal: reveal.ordinal_number.toString(),
+                  sat_rarity: satoshi.rarity,
+                  sat_coinbase_height: satoshi.blockHeight,
                 },
                 location: {
                   block_hash,
@@ -154,79 +158,40 @@ export class PgStore extends BasePgStore {
                   prev_offset: null,
                   value: reveal.inscription_output_value.toString(),
                   timestamp: event.timestamp,
-                  sat_ordinal: reveal.ordinal_number.toString(),
-                  sat_rarity: satoshi.rarity,
-                  sat_coinbase_height: satoshi.blockHeight,
                 },
               });
-              logger.info(
-                `PgStore cursed reveal #${reveal.inscription_number} (${reveal.inscription_id}) at block ${block_height}`
-              );
             }
             if (operation.inscription_transferred) {
               const transfer = operation.inscription_transferred;
               const satpoint = parseSatPoint(transfer.satpoint_post_transfer);
               const prevSatpoint = parseSatPoint(transfer.satpoint_pre_transfer);
-              const inscriptionId = await sql<{ id: string }[]>`
-                SELECT id FROM inscriptions WHERE genesis_id = ${transfer.inscription_id}
-              `;
-              if (inscriptionId.count) {
-                const inscription_id = parseInt(inscriptionId[0].id);
-                let sat_ordinal: string;
-                let sat_rarity: string;
-                let sat_coinbase_height: number;
-                if (transfer.ordinal_number) {
-                  const satoshi = new OrdinalSatoshi(transfer.ordinal_number);
-                  sat_ordinal = satoshi.ordinal.toString();
-                  sat_rarity = satoshi.rarity;
-                  sat_coinbase_height = satoshi.blockHeight;
-                } else {
-                  const genesis = await this.getInscriptionGenesis({
-                    genesis_id: transfer.inscription_id,
-                  });
-                  sat_ordinal = genesis?.sat_ordinal ?? '0';
-                  sat_rarity = genesis?.sat_rarity ?? 'common';
-                  sat_coinbase_height = parseInt(genesis?.sat_coinbase_height ?? '767430');
-                }
-                await this.insertInscriptionTransfer({
-                  inscription_id,
-                  location: {
-                    block_hash,
-                    block_height,
-                    tx_id,
-                    genesis_id: transfer.inscription_id,
-                    address: transfer.updated_address,
-                    output: `${satpoint.tx_id}:${satpoint.vout}`,
-                    offset: satpoint.offset ?? null,
-                    prev_output: `${prevSatpoint.tx_id}:${prevSatpoint.vout}`,
-                    prev_offset: prevSatpoint.offset ?? null,
-                    value: transfer.post_transfer_output_value
-                      ? transfer.post_transfer_output_value.toString()
-                      : null,
-                    timestamp: event.timestamp,
-                    // TODO: Store these fields in `inscriptions` instead of `locations`.
-                    sat_ordinal,
-                    sat_rarity,
-                    sat_coinbase_height,
-                  },
-                });
-                logger.info(
-                  `PgStore transfer (${transfer.inscription_id}) to output ${satpoint.tx_id}:${satpoint.vout} at block ${block_height}`
-                );
-              } else {
-                logger.warn(
-                  { block_height, genesis_id: transfer.inscription_id },
-                  `PgStore ignoring transfer for an inscription that does not exist (${transfer.inscription_id})`
-                );
-              }
+              await this.insertLocation({
+                location: {
+                  block_hash,
+                  block_height,
+                  tx_id,
+                  genesis_id: transfer.inscription_id,
+                  address: transfer.updated_address,
+                  output: `${satpoint.tx_id}:${satpoint.vout}`,
+                  offset: satpoint.offset ?? null,
+                  prev_output: `${prevSatpoint.tx_id}:${prevSatpoint.vout}`,
+                  prev_offset: prevSatpoint.offset ?? null,
+                  value: transfer.post_transfer_output_value
+                    ? transfer.post_transfer_output_value.toString()
+                    : null,
+                  timestamp: event.timestamp,
+                },
+              });
             }
           }
         }
+        updatedBlockHeightMin = Math.min(updatedBlockHeightMin, event.block_identifier.index);
       }
     });
     await this.refreshMaterializedView('chain_tip');
     // Skip expensive view refreshes if we're not streaming any live blocks yet.
     if (payload.chainhook.is_streaming_blocks) {
+      await this.normalizeInscriptionCount({ min_block_height: updatedBlockHeightMin });
       await this.refreshMaterializedView('inscription_count');
       await this.refreshMaterializedView('mime_type_counts');
       await this.refreshMaterializedView('sat_rarity_counts');
@@ -288,6 +253,16 @@ export class PgStore extends BasePgStore {
     return result[0].max.toString();
   }
 
+  async getInscriptionsPerBlockETag(): Promise<string> {
+    const result = await this.sql<{ block_hash: string; inscription_count: string }[]>`
+      SELECT block_hash, inscription_count
+      FROM inscriptions_per_block
+      ORDER BY block_height DESC
+      LIMIT 1
+    `;
+    return `${result[0].block_hash}:${result[0].inscription_count}`;
+  }
+
   async getInscriptionContent(
     args: InscriptionIdentifier
   ): Promise<DbInscriptionContent | undefined> {
@@ -309,13 +284,13 @@ export class PgStore extends BasePgStore {
     const result = await this.sql<{ etag: string }[]>`
       SELECT date_part('epoch', l.timestamp)::text AS etag
       FROM locations AS l
+      INNER JOIN current AS c ON l.id = c.location_id
       INNER JOIN inscriptions AS i ON l.inscription_id = i.id
       WHERE ${
         'genesis_id' in args
           ? this.sql`i.genesis_id = ${args.genesis_id}`
           : this.sql`i.number = ${args.number}`
       }
-      AND l.current = TRUE
     `;
     if (result.count > 0) {
       return result[0].etag;
@@ -335,10 +310,10 @@ export class PgStore extends BasePgStore {
       let orderBy = sql`gen.block_height`;
       switch (sort?.order_by) {
         case OrderBy.ordinal:
-          orderBy = sql`loc.sat_ordinal`;
+          orderBy = sql`i.sat_ordinal`;
           break;
         case OrderBy.rarity:
-          orderBy = sql`ARRAY_POSITION(ARRAY['common','uncommon','rare','epic','legendary','mythic'], loc.sat_rarity)`;
+          orderBy = sql`ARRAY_POSITION(ARRAY['common','uncommon','rare','epic','legendary','mythic'], i.sat_rarity)`;
           break;
       }
       // `ORDER` statement
@@ -360,6 +335,9 @@ export class PgStore extends BasePgStore {
           i.content_length,
           i.fee AS genesis_fee,
           i.curse_type,
+          i.sat_ordinal,
+          i.sat_rarity,
+          i.sat_coinbase_height,
           gen.block_height AS genesis_block_height,
           gen.block_hash AS genesis_block_hash,
           gen.tx_id AS genesis_tx_id,
@@ -369,11 +347,8 @@ export class PgStore extends BasePgStore {
           loc.address,
           loc.output,
           loc.offset,
-          loc.sat_ordinal,
-          loc.sat_rarity,
           loc.timestamp,
           loc.value,
-          loc.sat_coinbase_height,
           ${
             countType === DbInscriptionIndexResultCountType.custom
               ? sql`COUNT(*) OVER() as total`
@@ -410,12 +385,12 @@ export class PgStore extends BasePgStore {
           }
           ${
             filters?.from_sat_coinbase_height
-              ? sql`AND loc.sat_coinbase_height >= ${filters.from_sat_coinbase_height}`
+              ? sql`AND i.sat_coinbase_height >= ${filters.from_sat_coinbase_height}`
               : sql``
           }
           ${
             filters?.to_sat_coinbase_height
-              ? sql`AND loc.sat_coinbase_height <= ${filters.to_sat_coinbase_height}`
+              ? sql`AND i.sat_coinbase_height <= ${filters.to_sat_coinbase_height}`
               : sql``
           }
           ${
@@ -430,10 +405,10 @@ export class PgStore extends BasePgStore {
           }
           ${
             filters?.from_sat_ordinal
-              ? sql`AND loc.sat_ordinal >= ${filters.from_sat_ordinal}`
+              ? sql`AND i.sat_ordinal >= ${filters.from_sat_ordinal}`
               : sql``
           }
-          ${filters?.to_sat_ordinal ? sql`AND loc.sat_ordinal <= ${filters.to_sat_ordinal}` : sql``}
+          ${filters?.to_sat_ordinal ? sql`AND i.sat_ordinal <= ${filters.to_sat_ordinal}` : sql``}
           ${filters?.number?.length ? sql`AND i.number IN ${sql(filters.number)}` : sql``}
           ${filters?.from_number ? sql`AND i.number >= ${filters.from_number}` : sql``}
           ${filters?.to_number ? sql`AND i.number <= ${filters.to_number}` : sql``}
@@ -442,10 +417,10 @@ export class PgStore extends BasePgStore {
           ${filters?.output ? sql`AND loc.output = ${filters.output}` : sql``}
           ${
             filters?.sat_rarity?.length
-              ? sql`AND loc.sat_rarity IN ${sql(filters.sat_rarity)}`
+              ? sql`AND i.sat_rarity IN ${sql(filters.sat_rarity)}`
               : sql``
           }
-          ${filters?.sat_ordinal ? sql`AND loc.sat_ordinal = ${filters.sat_ordinal}` : sql``}
+          ${filters?.sat_ordinal ? sql`AND i.sat_ordinal = ${filters.sat_ordinal}` : sql``}
         ORDER BY ${orderBy} ${order}
         LIMIT ${page.limit}
         OFFSET ${page.offset}
@@ -467,20 +442,6 @@ export class PgStore extends BasePgStore {
         results: results ?? [],
       };
     });
-  }
-
-  async getInscriptionGenesis(args: { genesis_id: string }): Promise<DbLocation | undefined> {
-    const results = await this.sql<DbLocation[]>`
-      SELECT ${this.sql(LOCATIONS_COLUMNS.map(c => `l.${c}`))}
-      FROM locations AS l
-      INNER JOIN inscriptions AS i ON l.inscription_id = i.id
-      WHERE i.genesis_id = ${args.genesis_id}
-      ORDER BY l.block_height ASC
-      LIMIT 1
-    `;
-    if (results.count > 0) {
-      return results[0];
-    }
   }
 
   async getInscriptionLocations(
@@ -573,71 +534,46 @@ export class PgStore extends BasePgStore {
     }
   }
 
+  async getInscriptionCountPerBlock(
+    filters: DbInscriptionCountPerBlockFilters
+  ): Promise<DbInscriptionCountPerBlock[]> {
+    const fromCondition = filters.from_block_height
+      ? this.sql`block_height >= ${filters.from_block_height}`
+      : this.sql``;
+
+    const toCondition = filters.to_block_height
+      ? this.sql`block_height <= ${filters.to_block_height}`
+      : this.sql``;
+
+    const where =
+      filters.from_block_height && filters.to_block_height
+        ? this.sql`WHERE ${fromCondition} AND ${toCondition}`
+        : this.sql`WHERE ${fromCondition}${toCondition}`;
+
+    return await this.sql<DbInscriptionCountPerBlock[]>`
+      SELECT *
+      FROM inscriptions_per_block
+      ${filters.from_block_height || filters.to_block_height ? where : this.sql``}
+      ORDER BY block_height DESC
+      LIMIT 5000
+    `; // roughly 35 days of blocks, assuming 10 minute block times on a full database
+  }
+
   async refreshMaterializedView(viewName: string) {
     await this.sql`REFRESH MATERIALIZED VIEW ${
       isProdEnv ? this.sql`CONCURRENTLY` : this.sql``
     } ${this.sql(viewName)}`;
   }
 
-  private async insertInscriptionGenesis(args: {
+  private async insertInscription(args: {
     inscription: DbInscriptionInsert;
     location: DbLocationInsert;
   }): Promise<number | undefined> {
     let inscription_id: number | undefined;
     await this.sqlWriteTransaction(async sql => {
-      // Are we upserting?
-      const prevInscription = await sql<{ id: number }[]>`
+      const upsert = await sql<{ id: number }[]>`
         SELECT id FROM inscriptions WHERE number = ${args.inscription.number}
       `;
-      if (prevInscription.count !== 0) {
-        logger.info(
-          {
-            block_height: args.location.block_height,
-            genesis_id: args.inscription.genesis_id,
-          },
-          `PgStore upserting inscription genesis #${args.inscription.number}`
-        );
-      } else {
-        // Is this a sequential genesis insert?
-        if (args.inscription.number < 0) {
-          // Is it a cursed inscription?
-          const maxCursed = await this.getMaxCursedInscriptionNumber();
-          if (maxCursed !== undefined && maxCursed - 1 !== args.inscription.number) {
-            logger.info(
-              {
-                block_height: args.location.block_height,
-                genesis_id: args.inscription.genesis_id,
-              },
-              `PgStore inserting out-of-order cursed inscription genesis #${args.inscription.number}, current max is #${maxCursed}`
-            );
-          }
-        } else {
-          const maxNumber = await this.getMaxInscriptionNumber();
-          if (maxNumber !== undefined && maxNumber + 1 !== args.inscription.number) {
-            logger.info(
-              {
-                block_height: args.location.block_height,
-                genesis_id: args.inscription.genesis_id,
-              },
-              `PgStore inserting out-of-order inscription genesis #${args.inscription.number}, current max is #${maxNumber}`
-            );
-          }
-          // Is this a blessed inscription in a duplicate sat?
-          const dup = await sql<{ id: number }[]>`
-            SELECT id FROM locations WHERE sat_ordinal = ${args.location.sat_ordinal}
-          `;
-          if (dup.count > 0) {
-            logger.info(
-              {
-                block_height: args.location.block_height,
-                genesis_id: args.inscription.genesis_id,
-              },
-              `PgStore inserting duplicate blessed inscription in satoshi ${args.location.sat_ordinal}`
-            );
-          }
-        }
-      }
-
       const inscription = await sql<{ id: number }[]>`
         INSERT INTO inscriptions ${sql(args.inscription)}
         ON CONFLICT ON CONSTRAINT inscriptions_number_unique DO UPDATE SET
@@ -646,12 +582,16 @@ export class PgStore extends BasePgStore {
           content_type = EXCLUDED.content_type,
           content_length = EXCLUDED.content_length,
           content = EXCLUDED.content,
-          fee = EXCLUDED.fee
+          fee = EXCLUDED.fee,
+          sat_ordinal = EXCLUDED.sat_ordinal,
+          sat_rarity = EXCLUDED.sat_rarity,
+          sat_coinbase_height = EXCLUDED.sat_coinbase_height
         RETURNING id
       `;
       inscription_id = inscription[0].id;
       const location = {
         inscription_id,
+        genesis_id: args.location.genesis_id,
         block_height: args.location.block_height,
         block_hash: args.location.block_hash,
         tx_id: args.location.tx_id,
@@ -661,58 +601,69 @@ export class PgStore extends BasePgStore {
         prev_output: args.location.prev_output,
         prev_offset: args.location.prev_offset,
         value: args.location.value,
-        sat_ordinal: args.location.sat_ordinal,
-        sat_rarity: args.location.sat_rarity,
-        sat_coinbase_height: args.location.sat_coinbase_height,
         timestamp: sql`to_timestamp(${args.location.timestamp})`,
       };
       const locationRes = await sql<{ id: string }[]>`
         INSERT INTO locations ${sql(location)}
         ON CONFLICT ON CONSTRAINT locations_output_offset_unique DO UPDATE SET
           inscription_id = EXCLUDED.inscription_id,
+          genesis_id = EXCLUDED.genesis_id,
           block_height = EXCLUDED.block_height,
           block_hash = EXCLUDED.block_hash,
           tx_id = EXCLUDED.tx_id,
           address = EXCLUDED.address,
           value = EXCLUDED.value,
-          sat_ordinal = EXCLUDED.sat_ordinal,
-          sat_rarity = EXCLUDED.sat_rarity,
-          sat_coinbase_height = EXCLUDED.sat_coinbase_height,
           timestamp = EXCLUDED.timestamp
         RETURNING id
       `;
       await this.updateInscriptionLocationPointers({
         inscription_id,
+        genesis_id: args.inscription.genesis_id,
         location_id: locationRes[0].id,
         block_height: args.location.block_height,
       });
-      // const json = inscriptionContentToJson(args.inscription);
-      // if (json) {
-      //   const values = {
-      //     inscription_id,
-      //     p: json.p,
-      //     op: json.op,
-      //     content: json,
-      //   };
-      //   await sql`
-      //     INSERT INTO json_contents ${sql(values)}
-      //     ON CONFLICT ON CONSTRAINT json_contents_inscription_id_unique DO UPDATE SET
-      //       p = EXCLUDED.p,
-      //       op = EXCLUDED.op,
-      //       content = EXCLUDED.content
-      //   `;
-      // }
+      logger.info(
+        `PgStore${upsert.count > 0 ? ' upsert ' : ' '}reveal #${args.inscription.number} (${
+          args.location.genesis_id
+        }) at block ${args.location.block_height}`
+      );
     });
     return inscription_id;
   }
 
-  private async insertInscriptionTransfer(args: {
-    inscription_id: number;
-    location: DbLocationInsert;
-  }): Promise<void> {
+  private async insertLocation(args: { location: DbLocationInsert }): Promise<void> {
     await this.sqlWriteTransaction(async sql => {
+      // Does the inscription exist? Warn if it doesn't.
+      const genesis = await sql<{ id: number }[]>`
+        SELECT id FROM inscriptions WHERE genesis_id = ${args.location.genesis_id}
+      `;
+      if (genesis.count === 0) {
+        logger.warn(
+          `PgStore inserting transfer for missing inscription (${args.location.genesis_id}) at block ${args.location.block_height}`
+        );
+      }
+      const inscription_id = genesis.count > 0 ? genesis[0].id : null;
+      // Do we have the location from `prev_output`? Warn if we don't.
+      if (args.location.prev_output) {
+        const prev = await sql`
+          SELECT id FROM locations
+          WHERE genesis_id = ${args.location.genesis_id}
+            AND prev_output = ${args.location.prev_output}
+        `;
+        if (prev.count === 0) {
+          logger.warn(
+            `PgStore inserting transfer (${args.location.genesis_id}) superceding a missing prev_output ${args.location.prev_output} at block ${args.location.block_height}`
+          );
+        }
+      }
+      const upsert = await sql`
+        SELECT id FROM locations
+        WHERE genesis_id = ${args.location.genesis_id}
+        AND block_height = ${args.location.block_height}
+      `;
       const location = {
-        inscription_id: args.inscription_id,
+        inscription_id,
+        genesis_id: args.location.genesis_id,
         block_height: args.location.block_height,
         block_hash: args.location.block_hash,
         tx_id: args.location.tx_id,
@@ -722,83 +673,135 @@ export class PgStore extends BasePgStore {
         prev_output: args.location.prev_output,
         prev_offset: args.location.prev_offset,
         value: args.location.value,
-        sat_ordinal: args.location.sat_ordinal,
-        sat_rarity: args.location.sat_rarity,
-        sat_coinbase_height: args.location.sat_coinbase_height,
         timestamp: this.sql`to_timestamp(${args.location.timestamp})`,
       };
       const locationRes = await sql<{ id: string }[]>`
-        INSERT INTO locations ${this.sql(location)}
+        INSERT INTO locations ${sql(location)}
         ON CONFLICT ON CONSTRAINT locations_output_offset_unique DO UPDATE SET
           inscription_id = EXCLUDED.inscription_id,
+          genesis_id = EXCLUDED.genesis_id,
           block_height = EXCLUDED.block_height,
           block_hash = EXCLUDED.block_hash,
           tx_id = EXCLUDED.tx_id,
           address = EXCLUDED.address,
           value = EXCLUDED.value,
-          sat_ordinal = EXCLUDED.sat_ordinal,
-          sat_rarity = EXCLUDED.sat_rarity,
-          sat_coinbase_height = EXCLUDED.sat_coinbase_height,
           timestamp = EXCLUDED.timestamp
         RETURNING id
       `;
-      await this.updateInscriptionLocationPointers({
-        inscription_id: args.inscription_id,
-        location_id: locationRes[0].id,
-        block_height: args.location.block_height,
-      });
+      if (inscription_id) {
+        await this.updateInscriptionLocationPointers({
+          inscription_id,
+          genesis_id: args.location.genesis_id,
+          location_id: locationRes[0].id,
+          block_height: args.location.block_height,
+        });
+      }
+      logger.info(
+        `PgStore${upsert.count > 0 ? ' upsert ' : ' '}transfer (${
+          args.location.genesis_id
+        }) at block ${args.location.block_height}`
+      );
     });
   }
 
-  private async rollBackInscriptionGenesis(args: { genesis_id: string }): Promise<void> {
+  private async normalizeInscriptionCount(args: { min_block_height: number }): Promise<void> {
+    await this.sqlWriteTransaction(async sql => {
+      await sql`
+        DELETE FROM inscriptions_per_block
+        WHERE block_height >= ${args.min_block_height}
+      `;
+      // - gets highest total for a block < min_block_height
+      // - calculates new totals for all blocks >= min_block_height
+      // - inserts new totals
+      await sql`
+        WITH previous AS (
+          SELECT *
+          FROM inscriptions_per_block
+          WHERE block_height < ${args.min_block_height}
+          ORDER BY block_height DESC
+          LIMIT 1
+        ), updated_blocks AS (
+          SELECT
+            l.block_height,
+            MIN(l.block_hash),
+            COUNT(*) AS inscription_count,
+            COALESCE((SELECT previous.inscription_count_accum FROM previous), 0) + (SUM(COUNT(*)) OVER (ORDER BY l.block_height ASC)) AS inscription_count_accum,
+            MIN(l.timestamp)
+          FROM locations AS l
+          INNER JOIN genesis AS g ON g.location_id = l.id
+          WHERE l.block_height >= ${args.min_block_height}
+          GROUP BY l.block_height
+          ORDER BY l.block_height ASC
+        )
+        INSERT INTO inscriptions_per_block
+        SELECT * FROM updated_blocks
+        ON CONFLICT (block_height) DO UPDATE SET
+          block_hash = EXCLUDED.block_hash,
+          inscription_count = EXCLUDED.inscription_count,
+          inscription_count_accum = EXCLUDED.inscription_count_accum,
+          timestamp = EXCLUDED.timestamp;
+      `;
+    });
+  }
+
+  private async rollBackInscription(args: {
+    genesis_id: string;
+    number: number;
+    block_height: number;
+  }): Promise<void> {
     // This will cascade into dependent tables.
     await this.sql`DELETE FROM inscriptions WHERE genesis_id = ${args.genesis_id}`;
+    logger.info(
+      `PgStore rollback reveal #${args.number} (${args.genesis_id}) at block ${args.block_height}`
+    );
   }
 
-  private async rollBackInscriptionTransfer(args: {
+  private async rollBackLocation(args: {
     genesis_id: string;
     output: string;
-  }): Promise<number | undefined> {
-    let inscription_id: number | undefined;
-    await this.sqlWriteTransaction(async sql => {
-      const inscription = await sql<{ id: number }[]>`
-        SELECT id FROM inscriptions WHERE genesis_id = ${args.genesis_id}
-      `;
-      if (inscription.count === 0) {
-        logger.warn(
-          args,
-          `PgStore ignoring rollback for a transfer that does not exist (${args.genesis_id})`
-        );
-        return;
-      }
-      inscription_id = inscription[0].id;
-      await sql`
-        DELETE FROM locations
-        WHERE inscription_id = ${inscription_id} AND output = ${args.output}
-      `;
-    });
-    return inscription_id;
+    block_height: number;
+  }): Promise<void> {
+    await this.sql`
+      DELETE FROM locations
+      WHERE genesis_id = ${args.genesis_id} AND output = ${args.output}
+    `;
+    logger.info(
+      `PgStore rollback transfer (${args.genesis_id}) on output ${args.output} at block ${args.block_height}`
+    );
   }
 
   private async updateInscriptionLocationPointers(args: {
     inscription_id: number;
+    genesis_id: string;
     location_id: string;
     block_height: number;
   }): Promise<void> {
     await this.sqlWriteTransaction(async sql => {
+      // Update genesis and current location pointers for this inscription.
+      const pointer = {
+        inscription_id: args.inscription_id,
+        location_id: args.location_id,
+        block_height: args.block_height,
+      };
       await sql`
-        INSERT INTO genesis ${sql(args)}
+        INSERT INTO genesis ${sql(pointer)}
         ON CONFLICT ON CONSTRAINT genesis_inscription_id_unique DO UPDATE SET
           location_id = EXCLUDED.location_id,
           block_height = EXCLUDED.block_height
         WHERE EXCLUDED.block_height < genesis.block_height
       `;
       await sql`
-        INSERT INTO current ${sql(args)}
+        INSERT INTO current ${sql(pointer)}
         ON CONFLICT ON CONSTRAINT current_inscription_id_unique DO UPDATE SET
           location_id = EXCLUDED.location_id,
           block_height = EXCLUDED.block_height
         WHERE EXCLUDED.block_height > current.block_height
+      `;
+      // Backfill orphan locations for this inscription, if any.
+      await sql`
+        UPDATE locations
+        SET inscription_id = ${args.inscription_id}
+        WHERE genesis_id = ${args.genesis_id} AND inscription_id IS NULL
       `;
     });
   }