+++ tests/inscriptions.test.ts
@@ -307,7 +307,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'bdda0d240132bab2af7f797d1507beb1acab6ad43e2c0ef7f96291aea5cc3444:0:0',
             post_transfer_output_value: 9000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -357,7 +356,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'e3af144354367de58c675e987febcb49f17d6c19e645728b833fe95408feab85:0:0',
             post_transfer_output_value: 8000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -468,7 +466,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'bdda0d240132bab2af7f797d1507beb1acab6ad43e2c0ef7f96291aea5cc3444:0:0',
             post_transfer_output_value: 9000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -550,7 +547,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'bdda0d240132bab2af7f797d1507beb1acab6ad43e2c0ef7f96291aea5cc3444:0:0',
             post_transfer_output_value: 9000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -600,7 +596,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'e3af144354367de58c675e987febcb49f17d6c19e645728b833fe95408feab85:0:0',
             post_transfer_output_value: 8000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -687,7 +682,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'bdda0d240132bab2af7f797d1507beb1acab6ad43e2c0ef7f96291aea5cc3444:0:0',
             post_transfer_output_value: 9000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -743,7 +737,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'e3af144354367de58c675e987febcb49f17d6c19e645728b833fe95408feab85:0:0',
             post_transfer_output_value: 8000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -869,7 +862,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'bdda0d240132bab2af7f797d1507beb1acab6ad43e2c0ef7f96291aea5cc3444:0:0',
             post_transfer_output_value: 9000,
-            ordinal_number: null,
           })
           .transaction({
             hash: 'abe7deebd0c6bacc9b1ddd234f9442db0530180448e934f34b9cbf3d7e6d91cb',
@@ -882,7 +874,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               'abe7deebd0c6bacc9b1ddd234f9442db0530180448e934f34b9cbf3d7e6d91cb:0:0',
             post_transfer_output_value: 9000,
-            ordinal_number: null,
           })
           .build()
       );
@@ -968,7 +959,6 @@ describe('/inscriptions', () => {
             satpoint_post_transfer:
               '5cabafe04aaf98b1f325b0c3ffcbff904dbdb6f3d2e9e451102fda36f1056b5e:0:0',
             post_transfer_output_value: 8000,
-            ordinal_number: null,
           })
           .build()
       );